+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - dbt structure for data transformations
     - Loading from source.yml loads all data in stage
     - loading from stage_gas_staging loads just the last dtl loaded data (as it is a dbt full load, just the new data is loaded)
+    - adding country reporting table - for annual consumption data
     - Consider loading from stage_gas directly: should load everything again.
     - Consider using dbt incremental load